[Add] WorkSpaceListEntity 추가, GetWorkSpaceInfo 삭제
