+++ BranchTalk/Model/NetworkModel.swift
@@ -40,23 +40,6 @@ struct TokenResponse: Decodable {
     let accessToken: String
 }
 
-struct GetWorkSpaceInfo: Decodable {
-    let workspaceID: Int
-    let description: String?
-    let name, thumbnail: String
-    let ownerID: Int
-    let createdAt: String
-    
-    enum CodingKeys: String, CodingKey {
-        case workspaceID = "workspace_id"
-        case name, description, thumbnail
-        case ownerID = "owner_id"
-        case createdAt
-    }
-}
-
-typealias GetWorkSpaceList = [GetWorkSpaceInfo]
-
 struct MyInfo: Decodable {
     let userID, sesacCoin: Int
     let email, nickname, createdAt: String
